+++ tallyInc/meshConstruct.h
@@ -2,8 +2,8 @@
   CombLayer : MNCPX Input builder
  
  * File:   tallyInc/meshConstruct.h
-*
- * Copyright (c) 2004-2013 by Stuart Ansell
+ *
+ * Copyright (c) 2004-2015 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -48,7 +48,7 @@ class meshConstruct : virtual public basicConstruct
 
   void rectangleMesh(Simulation&,const int,const std::string&,
 		     const Geometry::Vec3D&,const Geometry::Vec3D&,
-		     const int*) const;
+		     const size_t*) const;
